finalized simulation script; even descritization among processors

+++ parallel_diffusion/plots.py
@@ -5,16 +5,16 @@ import glob
 import numpy as np
 
 procs = [2,4,8,16,32,64]
-xdivs = [2,4,8,16,32,64,128]
+xdivs = [2,4,8,16,32,64,128,192]
+# results = [1]
+delta_t = "plot_t301"
 results = [1,2,3]
-delta_t = "t301"
-if   delta_t == "t201": results = [1,2,3,4,5]
-elif delta_t == "t301": results = [1,2,3]
+
 
 path = os.path.join(os.getcwd())
 results_dir = f"{delta_t}_results_compiled"
 if not(os.path.isdir(results_dir)):
-    os.mkdir(results_dir)
+    os.makedirs(results_dir)
 
 
 # collect out file from each case
@@ -22,16 +22,22 @@ for r in results:
     i = 0
     for p in procs:
         for d in xdivs[i:]:
-            case_name = os.path.join(f"{delta_t}/results_{r}",f"p-{p}", f"divs-{d}")
+            case_name = os.path.join(f"results_{r}",f"p-{p}", f"divs-{d}")
             case_dir = os.path.join(path,case_name)
             new_out = f"out-{r}-{p}-{d}.txt"
-            shutil.copy(os.path.join(case_dir,"out.txt"), results_dir)
+            if (os.path.isfile(os.path.join(case_dir,"out.txt"))):
+                shutil.copy(os.path.join(case_dir,"out.txt"), results_dir)
+            else:
+                print(f"{new_out} does not exist")
             # shutil.copy(os.path.join(case_dir, new_out), os.path.join(case_dir, "out.txt"))
             if (os.path.isfile(os.path.join(results_dir, "out.txt"))):
                 shutil.move(os.path.join(results_dir, "out.txt"), os.path.join(results_dir,new_out))
+            else:
+                print(f"{new_out} skipped")
             
         i = i + 1
 
+# prepare times dictionary for each case.
 times = {}
 for r in results:
     i = 0
@@ -41,20 +47,21 @@ for r in results:
             times[case] = []
         i = i + 1
 
+# collect times for each case
 for r in results:
     i = 0
     for p in procs:
         for d in xdivs[i:]:
             out_file = os.path.join(results_dir,f"out-{r}-{p}-{d}.txt")
-            with open(out_file) as file:
-                case= f"{r}-{d}"
-                read = file.read()
-                index = read.find("time: ")
-                times[case].append(float(read[index:].split()[1]))
+            if (os.path.isfile(os.path.join(out_file))):
+                with open(out_file) as file:
+                    case= f"{r}-{d}"
+                    read = file.read()
+                    index = read.find("time: ")
+                    times[case].append(float(read[index:].split()[1]))
         i = i + 1
 
-
-plots_dir = f"{delta_t}_plot_results"
+plots_dir = f"{delta_t}"
 if not(os.path.isdir(plots_dir)):
     os.mkdir(plots_dir)
 
@@ -69,7 +76,7 @@ def print_plots(div):
     set_axis = 1
     for case in times.keys():
         if int(case.split("-")[1]) == div:
-            print(case, times[case])
+            # print(case, times[case])
             x = np.concatenate( (x,np.array( [times[case]])) , axis=set_axis)
             set_axis=0
             ax.set_xticks([i for i in range(2,2**len(times[case])+2,2)])
@@ -81,5 +88,8 @@ def print_plots(div):
     plt.legend()
     plt.savefig(os.path.join(plots_dir, f"{div}-plot"), dpi=300)
 
+
 for d in xdivs:
     print_plots(d)
+
+print(f"plots created at {results_dir}")
